Migrate CopyField to LitElement (#3302)
